requirements.txt
django-filter==23.1
Django==4.1.7
python-dotenv==1.0.0
drf_yasg==1.21.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5